Log unit test output interleaved with Phalcon system messages (debug).
git-svn-id: svn+ssh://svn.bmc.uu.se/cbe/openexam-php/branches/phalcon-mvc@1732 58b8aac4-d633-40de-91bf-eab09b0297ab

+++ tests/phalcon-mvc/support/Phalcon/LoggingCallback.php
@@ -110,6 +110,10 @@ class LoggingCallback extends FileAdapter
          */
         private function apply()
         {
+                if (!($stat = stat($this->options->logfile))) {
+                        return false;
+                }
+                
                 if ($this->options->truncate) {
                         unlink($this->options->logfile);
                 } elseif ($this->options->rotate) {
